Renamed to numa placement exporter
