+++ pkg/npe/cputopology/cputopology_test.go
