+++ pkg/npe/placement/placement.go
@@ -14,7 +14,7 @@ import (
 	"k8s.io/apimachinery/pkg/api/resource"
 	podresourcesv1 "k8s.io/kubelet/pkg/apis/podresources/v1"
 
-	"github.com/kai-scheduler/KAI-scheduler/pkg/numaagent/cputopology"
+	"github.com/kai-scheduler/KAI-scheduler/pkg/npe/cputopology"
 )
 
 const (
